+++ mapjs/src/argmapjs-utilities.js
@@ -1,3 +1,4 @@
+// TODO: switch to lodash and test
 const _ = require('underscore');
 
 idea_pp = function (idea, level = -1, key = []) {
@@ -38,3 +39,23 @@ ideas_pp = function (ideas, level = -1, keys = []) {
 }
 
 // ideas_pp(idea);
+
+// Two different ways of combining two theme files, see which one works best
+// order matters in both merge types, better results from having idea.theme as second argument.
+// 	Assume that value of second used where there is a clash.
+const mergeObjects = function (theme1 = themeProvider.default, theme2 = idea.theme, method = "lodash") {
+  switch (method) {
+
+    case "lodash":
+      const deep_merge = require('lodash/merge');
+      var merged_theme = deep_merge(theme1, theme2);
+      break;
+
+    default:
+      // QUESTION: This spread operator (...x) worked before: maybe debugger uses ES6 but webpack doesn't?
+      // var merged_theme = { ...theme1, ...theme1 };
+      var merged_theme = Object.assign({}, theme1, theme2);
+
+  }
+  return merged_theme
+};